Update menu.py (#280)

+++ menu.py
@@ -728,7 +728,7 @@ class SettingsMenu(FluentWindow):
         text_scale_factor.setText(str(float(conf.read_conf('General', 'scale')) * 100) + '%')  # 初始化缩放系数显示
 
         slider_scale_factor = self.adInterface.findChild(Slider, 'slider_scale_factor')
-        slider_scale_factor.setValue(float(conf.read_conf('General', 'scale')) * 100)
+        slider_scale_factor.setValue(int(float(conf.read_conf('General', 'scale')) * 100))
         slider_scale_factor.valueChanged.connect(
             lambda: (conf.write_conf('General', 'scale', str(slider_scale_factor.value() / 100)),
                      text_scale_factor.setText(str(slider_scale_factor.value()) + '%'))
